Move CallSiteProfile data class to own file
